Add free dates end point to contoller

+++ Modules/VisitSchedule/DentalClinic.VisitSchedule.Application/DTO/FreeDatesFilterDto.cs
@@ -2,7 +2,7 @@
 public class FreeDatesFilterDto
 {
     public long DoctorId { get; set; }
-    public long VisitTypeId { get; set; }
+    public int VisitTypeId { get; set; }
     public DateTime DateFrom { get; set; }
     public DateTime DateTo { get; set; }
 }
